Preparing Bricks for PR
- Cleaning the build from two separate binaries
- Adding code that runs in Bricks for testing
- C code is only user mode now
- Refactoring both Bricks and Tychools

+++ crates/bricks/src/gate_calls.rs
@@ -6,12 +6,12 @@ use crate::bricks_entry::current_frame;
 use crate::shared_buffer::bricks_write_ret_code;
 
 extern "C" {
-    fn asm_call_gate(capa_handle: &u64, args: &*const c_void) -> u32;
+    fn asm_call_gate(capa_handle: &u64, args: &*const c_void) -> u64;
 }
 
 // Gate call to return to untrusted part
-pub fn bricks_gate_call() -> u32 {
-    let mut result: u32 = FAILURE;
+pub fn bricks_gate_call() -> u64 {
+    let mut result: u64 = FAILURE;
     unsafe {
         if let Some(frame_curr) = &current_frame {
             result = asm_call_gate(&(frame_curr.handle), &(frame_curr.args));
